# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /alerts/article-metadata:
    post:
      operationId: fetchArticleMetadata
      tags: [alerts]
      summary: Read public article metadata from a URL
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArticleMetadataRequest"
      responses:
        "200":
          description: Extracted article metadata
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArticleMetadata"
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Article metadata could not be read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /alerts/scan:
    post:
      operationId: scanMarketAlerts
      tags: [alerts]
      summary: Scan tracked stocks and sectors for new market signals
      description: Reads the configured news and social sources and returns positive or negative items for tracked stocks and their sectors
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarketAlertScanRequest"
      responses:
        "200":
          description: Market intelligence alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MarketAlertScanResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    MarketAlertScanRequest:
      type: object
      properties:
        watchlist:
          type: array
          maxItems: 20
          items:
            $ref: "#/components/schemas/MarketAlertWatch"
        trackedArticles:
          type: array
          maxItems: 100
          items:
            $ref: "#/components/schemas/TrackedMarketArticle"
      required:
        - watchlist
    MarketAlertWatch:
      type: object
      properties:
        ticker:
          type: string
        companyName:
          type: string
          nullable: true
        sector:
          type: string
          nullable: true
      required:
        - ticker
    TrackedMarketArticle:
      type: object
      properties:
        ticker:
          type: string
        title:
          type: string
        url:
          type: string
        source:
          type: string
        publishedAt:
          type: string
        summary:
          type: string
          nullable: true
        articleType:
          $ref: "#/components/schemas/ArticleType"
      required:
        - ticker
        - title
        - url
        - source
        - publishedAt
    ArticleMetadataRequest:
      type: object
      properties:
        url:
          type: string
          format: uri
      required:
        - url
    ArticleType:
      type: string
      enum: ["earnings", "legal", "merger", "product", "leadership", "regulation", "analyst", "market", "other"]
    ArticleMetadata:
      type: object
      properties:
        title:
          type: string
        source:
          type: string
        publishedAt:
          type: string
        summary:
          type: string
        articleType:
          $ref: "#/components/schemas/ArticleType"
        fetchedAt:
          type: string
      required:
        - title
        - source
        - publishedAt
        - summary
        - articleType
        - fetchedAt
    MarketAlertScanResponse:
      type: object
      properties:
        alerts:
          type: array
          items:
            $ref: "#/components/schemas/MarketAlert"
        scannedTickers:
          type: integer
        scannedSectors:
          type: integer
        sources:
          type: array
          items:
            type: string
        checkedAt:
          type: string
      required:
        - alerts
        - scannedTickers
        - scannedSectors
        - sources
        - checkedAt
    MarketAlert:
      type: object
      properties:
        id:
          type: string
        subjectType:
          type: string
          enum: ["stock", "sector"]
        ticker:
          type: string
        subject:
          type: string
        title:
          type: string
        source:
          type: string
        url:
          type: string
        summary:
          type: string
        publishedAt:
          type: string
        sentiment:
          type: string
          enum: ["positive", "negative", "neutral"]
        articleType:
          $ref: "#/components/schemas/ArticleType"
        qualityScore:
          type: integer
          minimum: 0
          maximum: 100
        impactTitle:
          type: string
        impactSummary:
          type: string
        companyImpact:
          $ref: "#/components/schemas/NewsImpact"
        sectorImpact:
          $ref: "#/components/schemas/NewsImpact"
      required:
        - id
        - subjectType
        - ticker
        - subject
        - title
        - source
        - url
        - summary
        - publishedAt
        - sentiment
        - articleType
        - qualityScore
        - impactTitle
        - impactSummary
        - companyImpact
        - sectorImpact
    NewsImpactLabel:
      type: string
      enum: ["positive", "negative", "neutral", "unknown"]
    NewsImpactConfidence:
      type: string
      enum: ["high", "medium", "low", "unknown"]
    NewsImpact:
      type: object
      properties:
        label:
          $ref: "#/components/schemas/NewsImpactLabel"
        confidence:
          $ref: "#/components/schemas/NewsImpactConfidence"
        reason:
          type: string
      required:
        - label
        - confidence
        - reason
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message